+++ src/campaigns/services/brands.service.ts
@@ -7,17 +7,24 @@ import { CreateBrandDto, UpdateBrandDto } from '../dtos/brand.dtos';
 
 @Injectable()
 export class BrandsService {
-  constructor(@InjectRepository(Brand) private brandsRepo: Repository<Brand>) {}
+  constructor(@InjectRepository(Brand) private brandsRepo: Repository<Brand>) { }
 
   findAll() {
-    return this.brandsRepo.find();
+    //TODO: el limite de respuesta "take" deberia ser configurable
+    return this.brandsRepo.find({
+      where: {
+        status: true
+      },
+      take: 100
+    });
   }
 
   findOne(id: number) {
     //TODO: Actualizar para trajer todas las campanias de los branches asociacos a al brand buscado
     const campaign = this.brandsRepo.findOne({
       where: {
-        id
+        id,
+        status: true
       },
       relations: ['branches'],
     });
@@ -34,11 +41,18 @@ export class BrandsService {
 
   async update(id: number, changes: UpdateBrandDto) {
     const brand = await this.findOne(id);
+    if (!brand) {
+      throw new NotFoundException(`Brand #${id} not found`);
+    }
     this.brandsRepo.merge(brand, changes);
     return this.brandsRepo.save(brand);
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    const brand = await this.findOne(id);
+    if (!brand) {
+      throw new NotFoundException(`Brand #${id} not found`);
+    }
     return this.brandsRepo.delete(id);
   }
 }
