+++ src/campaigns/services/users.service.ts
@@ -9,19 +9,24 @@ import { CreateUserDto, UpdateUserDto } from '../dtos/user.dtos';
 export class UsersService {
   constructor(
     @InjectRepository(User) private userRepo: Repository<User>,
-  ) {}
+  ) { }
 
   findAll() {
-    return this.userRepo.find();
+    return this.userRepo.find({
+      where: {
+        status: true
+      },
+      take: 100
+    });
   }
 
   async findOne(id: number) {
-    const user = this.userRepo.findOne({
-      where:{
+    const user = await this.userRepo.findOne({
+      where: {
         id,
         status: true
       },
-      relations: ['cash_point_user'],
+      relations: ['campaignUsers'],
     });
     if (!user) {
       throw new NotFoundException(`User #${id} not found`);
@@ -29,17 +34,18 @@ export class UsersService {
     return user;
   }
 
-  async getByIdentification(identification: string){
+  async getByIdentification(identification: string) {
 
     const where: FindOptionsWhere<User> = {};
-          
+
     where.status = true;
     where.identification = identification;
-      
-    const user = this.userRepo.findOne({
+
+    const user = await this.userRepo.findOne({
       where: where,
-      relations: ['cash_point_user'],
+      relations: ['campaignUsers'],
     });
+
     if (!user) {
       throw new NotFoundException(`User with identification #${identification} not found`);
     }
@@ -53,11 +59,22 @@ export class UsersService {
 
   async update(id: number, changes: UpdateUserDto) {
     const user = await this.findOne(id);
+    if (!user) {
+      throw new NotFoundException(`User with id #${id} not found`);
+    }
+
     this.userRepo.merge(user, changes);
     return this.userRepo.save(user);
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+
+    const user = await this.findOne(id);
+
+    if (!user) {
+      throw new NotFoundException(`User with id #${id} not found`);
+    }
+
     return this.userRepo.delete(id);
   }
 }
